nexus-orchestration/pyproject.toml
[project]
name = "nexus-orchestration"
version = "0.1.0"
description = "Nexus Temporal workers — orchestrates expense audit and RAG chatbot workflows"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "temporalio>=1.8",
    "motor>=3.5",
    "redis>=5.0",
    "python-ulid>=3.0",
    "pydantic>=2.6",
    "pydantic-settings>=2.2",
    "structlog>=24.1",
    "rapidfuzz>=3.9",
    "boto3>=1.34",
    # Vector Search + SQL are opt-in at runtime via FAKE_VECTOR_SEARCH /
    # FAKE_SQL_SEARCH. Ship the clients in the base image so we can flip
    # the flag without rebuild.
    "databricks-vectorsearch>=0.50",
    "databricks-sql-connector>=3.3",
    # Phase E.3 — OpenTelemetry breadcrumb tracing (ADOT collector → X-Ray)
    "opentelemetry-api>=1.29",
    "opentelemetry-sdk>=1.29",
    "opentelemetry-exporter-otlp-proto-grpc>=1.29",
    "opentelemetry-instrumentation-botocore>=0.50b0",
    "opentelemetry-instrumentation-pymongo>=0.50b0",
    "opentelemetry-instrumentation-redis>=0.50b0",
    "opentelemetry-instrumentation-grpc>=0.50b0",
    "opentelemetry-propagator-aws-xray>=1.0.2",
    "opentelemetry-sdk-extension-aws>=2.0.2",
]

[project.optional-dependencies]
prod = [
    "anthropic>=0.34",
    "prometheus-client>=0.20",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "mongomock-motor>=0.0.30",
    "fakeredis>=2.23",
    "ruff>=0.6",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/nexus_orchestration"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
